Added the `has()` method to the HeadersInterface.
Fixes #83
